+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   * (theimo1221) Add option to run specific ac's regardless of energy level
   * (theimo1221) Add option to disable dachs altogether
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Enforce warm water pump to be off if the temperature is too high
+
 ## 3.2.20 (2025-11-07)
 
 * (theimo1221) Increase Block duration to 15 minutes